extract search context logic into new SearchContext.tsx
